Add more test and fix previous merge problem
