--- src/__tests__/profile.test.js
@@ -1,18 +1,0 @@
-import React from 'react';
-import renderer from 'react-test-renderer';
-import { Provider } from 'react-redux';
-import store from '../redux/configureStore';
-import Profile from '../components/Profile';
-
-describe('Group of Profile tests', () => {
-  test('Test if Profile page is rendering correctly', () => {
-    const profile = renderer
-      .create(
-        <Provider store={store}>
-          <Profile />
-        </Provider>,
-      )
-      .toJSON();
-    expect(profile).toMatchSnapshot();
-  });
-});
